+++ actionview/lib/action_view/helpers/rendering_helper.rb
@@ -12,6 +12,8 @@ module ActionView
       # * <tt>:file</tt> - Renders an explicit template file (this used to be the old default), add :locals to pass in those.
       # * <tt>:inline</tt> - Renders an inline template similar to how it's done in the controller.
       # * <tt>:text</tt> - Renders the text passed in out.
+      # * <tt>:plain</tt> - Renders the text passed in out. Setting the content
+      # type as <tt>text/plain</tt>.
       # * <tt>:body</tt> - Renders the text passed in, and does not set content
       # type in the response.
       #